Change class name; remove unnecessary methods and properties
